+++ src/collatz/StepFactory.java
@@ -2,12 +2,12 @@ package collatz;
 
 public class StepFactory {
 
-    public static Step create(int number) {
+    public static Step create(long number) {
         if (isEven(number)) return new EvenStep(number);
         else return new OddStep(number);
     }
 
-    private static boolean isEven(int number) {
+    private static boolean isEven(long number) {
         return number % 2 == 0;
     }
 }
